+++ app/lib/filesharing/dialog/course_tile.dart
@@ -12,7 +12,7 @@ import 'package:flutter/material.dart';
 import 'package:group_domain_models/group_domain_models.dart';
 import 'package:sharezone/groups/group_join/group_join_page.dart';
 import 'package:sharezone/groups/group_permission.dart';
-import 'package:sharezone/groups/src/pages/shared/course_template_list.dart';
+import 'package:sharezone/groups/src/pages/shared/course_template_page.dart';
 import 'package:sharezone/main/application_bloc.dart';
 import 'package:sharezone/util/api.dart';
 import 'package:sharezone_widgets/sharezone_widgets.dart';
