+++ app/lib/groups/src/pages/course/group_page.dart
@@ -10,7 +10,7 @@ import 'package:bloc_provider/bloc_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_staggered_animations/flutter_staggered_animations.dart';
 import 'package:group_domain_models/group_domain_models.dart';
-import 'package:sharezone/groups/src/pages/shared/course_template_list.dart';
+import 'package:sharezone/groups/src/pages/shared/course_template_page.dart';
 import 'package:sharezone/main/application_bloc.dart';
 import 'package:sharezone/groups/analytics/group_analytics.dart';
 import 'package:sharezone/groups/group_join/group_join_page.dart';
