+++ app/lib/groups/src/pages/school_class/school_class_details/school_class_course_list.dart
@@ -10,7 +10,7 @@ import 'package:bloc_provider/bloc_provider.dart';
 import 'package:common_domain_models/common_domain_models.dart';
 import 'package:flutter/material.dart';
 import 'package:group_domain_models/group_domain_models.dart';
-import 'package:sharezone/groups/src/pages/shared/course_template_list.dart';
+import 'package:sharezone/groups/src/pages/shared/course_template_page.dart';
 import 'package:sharezone/main/application_bloc.dart';
 import 'package:sharezone/groups/group_permission.dart';
 import 'package:sharezone/groups/src/pages/course/course_card.dart';
