+++ app/lib/groups/src/pages/shared/course_template_page.dart
