+++ app/lib/timetable/src/widgets/timetable_day_view.dart
@@ -11,7 +11,7 @@ import 'package:date/date.dart';
 import 'package:date/weektype.dart';
 import 'package:flutter/material.dart';
 import 'package:group_domain_models/group_domain_models.dart';
-import 'package:sharezone/groups/src/pages/shared/course_template_list.dart';
+import 'package:sharezone/groups/src/pages/shared/course_template_page.dart';
 import 'package:sharezone/main/application_bloc.dart';
 import 'package:sharezone/calendrical_events/models/calendrical_event.dart';
 import 'package:sharezone/groups/group_join/group_join_page.dart';
